fix: remove document and move script to app
