--- pages/_document.tsx
@@ -1,47 +1,0 @@
-/* eslint-disable max-len */
-/* eslint-disable require-jsdoc */
-import React from 'react';
-import Document, {Html, Head, Main, NextScript} from 'next/document';
-import Script from 'next/script';
-
-class MyDocument extends Document {
-  render() {
-    return (
-      <Html>
-        <Head>
-          <Script src="https://developers.kakao.com/sdk/js/kakao.min.js"/>
-          <Script
-            async
-            src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}`}
-          />
-          <Script
-            dangerouslySetInnerHTML={{
-              __html: `
-                window.dataLayer = window.dataLayer || [];
-                function gtag(){dataLayer.push(arguments);}
-                gtag('js', new Date());
-                gtag('config', '${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}', {
-                  page_path: window.location.pathname,
-                });
-              `,
-            }}
-          />
-          <Script id="adsense-id" async
-            onError={(e) => {
-              console.error('Script failed to load', e);
-            }}
-            strategy="afterInteractive"
-            src={`https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=${process.env.NEXT_PUBLIC_GOOGLE_ADSENSE}`}
-            crossOrigin="anonymous"
-          />
-        </Head>
-        <body>
-          <Main />
-          <NextScript />
-        </body>
-      </Html>
-    );
-  }
-}
-
-export default MyDocument;
